collector/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
gtfs-realtime-bindings>=1.0.0
protobuf>=4.21.0
pytest>=7.0.0
